feat(Container): import semantic component and add overrides
