--- src/components/general-widgets/Container/Readme.md
@@ -1,5 +1,0 @@
-```jsx
- <Container>
-   <OwnerLogin />
- </Container>
-```
