+++ src/styles/semantic/semantic.less
@@ -17,6 +17,7 @@
 & { @import "~semantic-ui-less/definitions/globals/site.less"; }
 
 /* Elements */
+& { @import "~semantic-ui-less/definitions/elements/container.less"; }
 & { @import "~semantic-ui-less/definitions/elements/button.less"; }
 & { @import "~semantic-ui-less/definitions/elements/divider.less"; }
 & { @import "~semantic-ui-less/definitions/elements/flag.less"; }
