+++ src/styles/semantic/theme.config
@@ -23,7 +23,7 @@
 
 /* Elements */
 @button     : 'livingstone';
-@container  : 'default';
+@container  : 'livingstone';
 @divider    : 'default';
 @flag       : 'default';
 @header     : 'livingstone';
